// src/store/mutation-types.ts
export enum MutationTypes {
    addWeatherListItem = "addWeatherListItem",
    clearWeatherList = "clearWeatherList",

    clearSuggest = "clearSuggest",
    addSuggestList = "addSuggestList",

    addSaveListItem = "addSaveListItem",
    removeSaveListItem = "removeSaveListItem",
    changePositionSaveListItem = "changePositionSaveListItem",
    clearSaveList = "clearSaveList",

    addUpdatingRequest = "addUpdatingRequest",
    removeUpdatingRequest = "removeUpdatingRequest",
}